+++ src/Symfony/Component/Dotenv/Tests/Command/DotenvDumpCommandTest.php
@@ -12,12 +12,16 @@
 namespace Symfony\Component\Dotenv\Tests\Command;
 
 use PHPUnit\Framework\TestCase;
+use Symfony\Bridge\PhpUnit\ExpectDeprecationTrait;
 use Symfony\Component\Console\Application;
 use Symfony\Component\Console\Tester\CommandTester;
 use Symfony\Component\Dotenv\Command\DotenvDumpCommand;
+use Symfony\Component\Dotenv\Dotenv;
 
 class DotenvDumpCommandTest extends TestCase
 {
+    use ExpectDeprecationTrait;
+
     protected function setUp(): void
     {
         file_put_contents(__DIR__.'/.env', <<<EOF
@@ -37,6 +41,7 @@ EOF
         @unlink(__DIR__.'/.env');
         @unlink(__DIR__.'/.env.local');
         @unlink(__DIR__.'/.env.local.php');
+        @unlink(__DIR__.'/.env.dumped.php');
         @unlink(__DIR__.'/composer.json');
     }
 
@@ -45,11 +50,12 @@ EOF
         $command = $this->createCommand();
         $command->execute([
             'env' => 'test',
+            '--preferPhpFilesAndChangeDumpName' => null,
         ]);
 
-        $this->assertFileExists(__DIR__.'/.env.local.php');
+        $this->assertFileExists(__DIR__.'/.env.dumped.php');
 
-        $vars = require __DIR__.'/.env.local.php';
+        $vars = require __DIR__.'/.env.dumped.php';
         $this->assertSame([
             'APP_ENV' => 'test',
             'APP_SECRET' => 'abc123',
@@ -62,11 +68,12 @@ EOF
         $command->execute([
             'env' => 'test',
             '--empty' => true,
+            '--preferPhpFilesAndChangeDumpName' => null,
         ]);
 
-        $this->assertFileExists(__DIR__.'/.env.local.php');
+        $this->assertFileExists(__DIR__.'/.env.dumped.php');
 
-        $vars = require __DIR__.'/.env.local.php';
+        $vars = require __DIR__.'/.env.dumped.php';
         $this->assertSame(['APP_ENV' => 'test'], $vars);
     }
 
@@ -80,16 +87,30 @@ EOF
         $command = $this->createCommand();
         $command->execute([
             'env' => 'test',
+            '--preferPhpFilesAndChangeDumpName' => null,
         ]);
 
-        $this->assertFileExists(__DIR__.'/.env.local.php');
+        $this->assertFileExists(__DIR__.'/.env.dumped.php');
 
-        $vars = require __DIR__.'/.env.local.php';
+        $vars = require __DIR__.'/.env.dumped.php';
         $this->assertSame([
             'APP_ENV' => 'test',
             'APP_SECRET' => 'abc123',
             'APP_LOCAL' => 'yes',
         ], $vars);
+    }
+
+    /**
+     * @group legacy
+     */
+    public function testDeprecationForNotPreferringPhpFiles()
+    {
+        $this->expectDeprecation('Since symfony/dotenv 7.2: Not using --preferPhpFilesAndChangeDumpName is deprecated.');
+        $this->expectDeprecation('Since symfony/dotenv 7.2: Setting $preferPhpFilesAndChangeDumpName to false is deprecated.');
+        $command = $this->createCommand();
+        $command->execute([
+            'env' => 'test',
+        ]);
     }
 
     private function createCommand(): CommandTester